livecheck: clarify livecheckable language
Formulae, casks, and resources have a `#livecheckable?` method that
indicates whether they contain a `livecheck` block. This is intended
to be read as "has a livecheckable?", not "is livecheckable?".
Unfortunately, correct understanding of this method's behavior relies
on historical knowledge that few people possess, so this is often
confusing to anyone who hasn't been working on livecheck since 2020.

In the olden days, a "livecheckable" was a Ruby file containing a
`livecheck` block (originally a hash) with a filename that
corresponded to a related formula. The `livecheck` blocks in
livecheckable files were integrated into their respective formulae in
August 2020, so [first-party] livecheckables ceased to exist at that
time. From that point forward, we simply referred to these as
`livecheck` blocks.

With that in mind, this clarifies the situation by replacing
"livecheckable" language with "livecheck block". This includes
renaming `#livecheckable?` to `#livecheck_block?`, replacing usage of
"livecheckable" as a noun with "`livecheck` block", replacing
"livecheckable" as a boolean with "has_livecheck_block", and replacing
incorrect usage of "livecheckable" as an adjective with "checkable".